requirements.txt
flask
flask-sock
openai
bcrypt
python-dotenv
fastapi
uvicorn